[trace] fix race when child and parent send is called at the same time
Fixes https://github.com/honeycombio/beeline-go/issues/38  - if `Send` is called for a span and its parent at the same time, a race can occur when accesing "isSent". This adds a lock to `Send`, which sets `isSent` and protects reads with read locks. This also effectively prevents `Send` from executing multiple times for the same span.