Allow override of libhoney client
We introduced client objects in https://github.com/honeycombio/libhoney-go/pull/40 - now we want to enable the beeline to use the non-default libhoney client and enable better customization of the client.

This required some refactoring, as the `beeline` module and the `trace` module both shared the `libhoney` global state, and don't share any state within the library. To make this work, I pulled client initialization into its own module, and exposed the client methods as module methods. Since libhoney's `NewClient` returns an error, but beeline's `Init` does not, we don't have a great way to handle failed initializations, so the wrapper methods in the `client` module protect agains the case that client initialization fails and returns us a nil client.