[sqlx] functions with context should use spans; this one was skipped (#76)
All the functions that get context are supposed to use that context to join the events created to existing traces as spans. Functions that don't have context use the event-version of the instrumentation instead.

This instances of the QueryxContext function accidentally used the event version instead of the span version. This change fixes that error.