Generalize PropagationContext and Related Methods (#111)
Generalize propagation package in prep for additional formats.

The propagation package was very coupled to the Honeycomb trace context header format. We plan to add support for other formats in the future. This change prepares the beeline for new formats that can be converted to/from PropagationContext instances. Also adding NewTraceFromPropagationContext function to trace package. This adds NewTraceFromPropagationContext and NewTraceFromSerializedHeaders functions to the trace package. NewTrace wraps calls to the latter. This is just a refactoring change. In a future PR, we'll use the NewTraceFromPropagationContext to create traces from data parsed in a variety of header formats.