Disable validation check on Unmarshal for Honeycomb headers (#125)
The IsValid() function on propagation.PropagationContext is intended to
validate PropagationContext objects that are either created from
incoming trace context headers, or used to generate outgoing trace
context headers. In either case, the context had better contain a
ParentID or it is not valid.

There is a use case, however, with Honeycomb headers, where a user of
older beeline versions would create a PropagationContext (aliased as
Propagation) manually, and then serialize it as a Honeycomb header to
create a new trace.Trace. In this case, the PropagationContext would not
necessarily have a TraceID or a ParentId, and so it would fail IsValid.