Protect against potential data race in working with propagation context (#131)
There are two ways to get propagation headers out of the go beeline. The function to serialize headers (that returns the beeline header format) and the newer request for a PropagationContext object which can then be serialized into custom formats using the wrapper's propagation hooks.

Unfortunately the latter form does not protect the trace level fields map from concurrent access. The effect is that if one gets a propagation context object and simultaneously adds a trace level field, a data race on that map access can occur.

The first commit in this PR adds a test that confirms this race can happen. Switching your git checkout to f22a86c and running go test ./... -race will spit out the dreaded WARNING: DATA RACE message. Moving back to past the second commit in this PR and re-running the tests will show that moving the portion that works on the trace level fields into the Trace-attached method and making a copy of the fields protects against that race.

The interface for calling either of these methods is unchanged.