Wrappers: add cfg to disable client spans
Add a new `DisableClientSpans` config parameter to the hnynethttp and
hnygrpc wrappers. By default (when set to the Go zero-value, or when
created using the default wrapper constructors), this has no effect.
When specifically set to `true`, this config parameter suppresses the
creation of new spans by the client wrappers. The trace propagation
metadata is still sent with the request, but the span ID is the ID of
the current span, not the ID of a newly-created span.

This allows users to decide whether or not they care about client spans
independently of whether they care about trace propagation. Since client
spans can nearly double the total number of spans sent by a microservice
architecture, this can be an important knob to tune.