[sql instrumentation] set name field to db call, not caller
To keep with the data model of the name of the span representing the thing that it's doing, using the caller as the name is the wrong choice; it should be the db call being run. Keep db.caller as a separate field, useful when invoked, but use `Get` or `GetContext` and so on as the name of the span.