Add more request header information to GRPC handler spans (#341)
- Adds the following fields to a GRPC handler span:
  - request.header.x_forwarded_for
  - request.header.x_forwarded_proto
  - request.remote_addr

request.remote_addr is retrieved from the GRPC peer which may not
be the origin client IP when a proxy is in the middle. The field name
used is the equivalent from HTTP instrumentation.

- Refactored the existing if-chain to a loop of headers to check
  and add to span if present.

The list of request headers to put on the span is captured in a
package variable. It's the closest to a constant map we've got
before getting into Go shenanigans.

Co-authored-by: Kent Quirk <kentquirk@gmail.com>