bring back local variable
We have concluded that calling functions as parameters to functions is
not as easy to read as local variable assignment and using the variable
as the parameter. Benchmarking indicates that there is no difference in
allocations between assigning a local variable in AddField() and having
a function return a value that is immediately used as a parameter still
within the scope of AddField().